doc(odata2model): refactored documentation
